refactoring code, move each chart into separate js file
